+++ orix/crystal_map/crystal_map.py
@@ -731,7 +731,7 @@ class CrystalMap:
                 # Fill in orientations per phase
                 # TODO: Consider whether orientations should be calculated
                 #  upon loading
-                for i, phase in self.phases_in_data:
+                for i, _ in self.phases_in_data:
                     phase_mask = (self._phase_id == i) * self.is_in_data
                     phase_mask_in_data = self.phase_id == i
                     array[phase_mask] = self[phase_mask_in_data].orientations.to_euler()