+++ orix/tests/io/test_h5ebsd.py
@@ -16,11 +16,21 @@
 # You should have received a copy of the GNU General Public License
 # along with orix.  If not, see <http://www.gnu.org/licenses/>.
 
-from orix.io.plugins import ang, bruker_h5ebsd, emsoft_h5ebsd, orix_hdf5
+from h5py import File
 
-plugin_list = [
-    ang,
-    bruker_h5ebsd,
-    emsoft_h5ebsd,
-    orix_hdf5,
-]
+from orix.io import save
+from orix.io.plugins._h5ebsd import hdf5group2dict
+
+
+class TestH5ebsd:
+    def test_hdf5group2dict_update_dict(self, temp_file_path, crystal_map):
+        """Can read datasets from an HDF5 file into an existing
+        dictionary.
+        """
+        save(temp_file_path, crystal_map)
+        with File(temp_file_path, mode="r") as f:
+            this_dict = {"hello": "there"}
+            this_dict = hdf5group2dict(f["crystal_map"], dictionary=this_dict)
+            assert this_dict["hello"] == "there"
+            assert this_dict["data"] == f["crystal_map/data"]
+            assert this_dict["header"] == f["crystal_map/header"]