+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Added
   Rodrigues fundamental zone, without having to set it up via Matplotlib.
 - Method Object3d.get_random_sample(), inherited by all 3D objects, returning a new
   flattened instance with elements drawn randomly from the original instance.
+- Reading of a crystal map from orientation data in Bruker's HDF5 file format.
 
 2021-05-23 - version 0.6.0
 ==========================